Hakeem-Rag/src/models/db_schemes/asset.py
from email.policy import default
from operator import ge
from pydantic import BaseModel, Field
from typing import Optional
from bson.objectid import ObjectId
from datetime import datetime

class Asset(BaseModel):
    id: Optional[ObjectId] = Field(None, alias="_id")
    asset_project_id: ObjectId
    asset_type: str = Field(..., min_length=1) #url, text ...
    asset_name: str = Field(..., min_length=1) # file Id
    asset_size: int = Field(ge=0, default=None) # file size (optional)
    asset_config: dict = Field(default=None)
    asset_pushed_at: datetime = Field(default=datetime.now())

    class Config:
        arbitrary_types_allowed = True

    @classmethod
    def get_indexes(cls):
        return [
            {
                "key": [
                    ("asset_project_id", 1)
                ],
                "name": "asset_project_id_index_1",
                "unique": False
            },
             {
                "key": [
                    ("asset_project_id", 1),
                    ("asset_name", 1)
                ],
                "name": "asset_project_id_name_index_1",
                "unique": True
            }
        ]
