Hakeem-Rag/src/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.31.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings==2.5.2
aiofiles==24.1.0
ibm-watsonx-ai==1.1.20
langchain==0.3.7
langchain-community==0.3.5
PyMuPDF==1.24.13
motor==3.6.0
pytorch==1.0.2
torchvision==0.20.1
torchaudio==2.5.1
sentence-transformers==3.2.1